tests ready + flake8 + bandit + js in html + csrf protection

+++ geocoder/geocoding/download.py
@@ -13,7 +13,8 @@ from loguru import logger
 from urllib3 import disable_warnings
 from urllib3.exceptions import InsecureRequestWarning
 
-from geocoder.geocoding.datapaths import here
+from geocoder.geocoding.datapaths import here, database
+from geocoder.geocoding import DEBUG
 
 SSL_VERIFICATION = os.environ.get("SSL_VERIFICATION", False)
 
@@ -27,23 +28,26 @@ ban_dpt_file_name = [filegz[:-3] for filegz in ban_dpt_gz_file_name]
 content_folder_path = os.path.join(here, 'content')
 server_content_file_name = os.path.join(content_folder_path, 'server_content_v2.txt')
 local_content_file_name = os.path.join(content_folder_path, 'local_content_v2.txt')
-dpt_list = ["01", "02", "03", "04", "05", "06", "07", "08", "09",
-            "10", "11", "12", "13", "14", "15", "16", "17", "18", "19",
-            "2A", "2B", "21", "22", "23", "24", "25", "26", "27", "28", "29",
-            "30", "31", "32", "33", "34", "35", "36", "37", "38", "39",
-            "40", "41", "42", "43", "44", "45", "46", "47", "48", "49",
-            "50", "51", "52", "53", "54", "55", "56", "57", "58", "59",
-            "60", "61", "62", "63", "64", "65", "66", "67", "68", "69",
-            "70", "71", "72", "73", "74", "75", "76", "77", "78", "79",
-            "80", "81", "82", "83", "84", "85", "86", "87", "88", "89",
-            "90", "91", "92", "93", "94", "95",
-            "971", "972", "973", "974", "975", "976", "977", "978", "984", "986", "987", "988", "989"]
+if DEBUG:
+    dpt_list = ["01"]
+else:  # pragma: no cover
+    dpt_list = ["01", "02", "03", "04", "05", "06", "07", "08", "09",
+                "10", "11", "12", "13", "14", "15", "16", "17", "18", "19",
+                "2A", "2B", "21", "22", "23", "24", "25", "26", "27", "28", "29",
+                "30", "31", "32", "33", "34", "35", "36", "37", "38", "39",
+                "40", "41", "42", "43", "44", "45", "46", "47", "48", "49",
+                "50", "51", "52", "53", "54", "55", "56", "57", "58", "59",
+                "60", "61", "62", "63", "64", "65", "66", "67", "68", "69",
+                "70", "71", "72", "73", "74", "75", "76", "77", "78", "79",
+                "80", "81", "82", "83", "84", "85", "86", "87", "88", "89",
+                "90", "91", "92", "93", "94", "95",
+                "971", "972", "973", "974", "975", "976", "977", "978", "984", "986", "987", "988", "989"]
 
 
 def completion_bar(msg, fraction):
     percent = int(100 * fraction)
     size = int(50 * fraction)
-    sys.stdout.write("\r%-16s: %3d%%[%s%s]" %
+    sys.stdout.write("\r%-16s: %3d%%[%s%s]\n" %
                      (msg, percent, '=' * size, ' ' * (50 - size)))
     sys.stdout.flush()
 
@@ -65,10 +69,10 @@ def update_ban_file(url, file_name):
         msg = 'Unable to join BAN address website ({}).'.format(url)
         r = requests.get(url, verify=SSL_VERIFICATION)
         msg += ' Status error code: {}'.format(r.status_code)
-        if r.status_code != 200:
+        if r.status_code != 200:  # pragma: no cover
             raise ConnectionError(msg)
     except (requests.exceptions.ConnectionError, requests.exceptions.SSLError, requests.exceptions.HTTPError):
-        raise ConnectionError(msg)
+        raise ConnectionError(msg)  # pragma: no cover
 
     with open(file_name, 'w') as file:
         file.write(r.text)
@@ -88,7 +92,7 @@ def need_to_download():
     else:
         update_server_content_file()
 
-        if md5(server_content_file_name) == md5(local_content_file_name):
+        if md5(server_content_file_name) == md5(local_content_file_name) and os.path.exists(database):
             logger.info('BAN database is already up to date. No need to download it again.')
             os.remove(server_content_file_name)
             return False
@@ -101,7 +105,7 @@ def download_ban_dpt_file(ban_dpt_file_name):
         response = requests.get(ban_url.format(ban_dpt_file_name), stream=True,
                                 verify=SSL_VERIFICATION)
 
-        if not response.ok:
+        if not response.ok:  # pragma: no cover
             logger.info('Download {} unsuccessful: bad response'.format(ban_dpt_file_name))
             return False
 
@@ -110,7 +114,7 @@ def download_ban_dpt_file(ban_dpt_file_name):
             ban_dpt_file.write(block)
             done += len(block)
 
-    if done != total_size:
+    if done != total_size:  # pragma: no cover
         logger.error('Download {} unsuccessful: incomplete'.format(ban_dpt_file_name))
         return False
 
@@ -151,7 +155,7 @@ def decompress():
         # Certifies the existence of the subdirectory.
         for ban_dpt_gz_file_name_type in ban_dpt_gz_file_name:
             dpt_gz_file_path = os.path.join(raw_data_folder_path, ban_dpt_gz_file_name_type.format(dpt))
-            if not os.path.isfile(dpt_gz_file_path):
+            if not os.path.isfile(dpt_gz_file_path):  # pragma: no cover
                 logger.error('Decompression unsuccessful: nonexistent file {}'.format(dpt_gz_file_path))
                 count += 1
 
@@ -164,7 +168,7 @@ def decompress():
 
                 remove_file(dpt_gz_file_path)
 
-    if count:
+    if count:  # pragma: no cover
         return False
     return True
 
@@ -181,4 +185,7 @@ def remove_downloaded_raw_ban_files():
     for dpt in dpt_list:
         for ban_dpt_gz_file_name_type in ban_dpt_gz_file_name:
             remove_file(os.path.join(raw_data_folder_path, ban_dpt_gz_file_name_type.format(dpt)))
+        for ban_dpt_gz_file_name_type in ban_dpt_file_name:
+            remove_file(os.path.join(raw_data_folder_path, ban_dpt_gz_file_name_type.format(dpt)))
+    shutil.rmtree(raw_data_folder_path)
     return True
